requirements.txt
numpy>=1.18.0
pytest>=8.3.5
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
